+++ src/app/providers/remote-read/deck.spec.ts
@@ -19,7 +19,10 @@ vi.mock("firebase/firestore", () => ({
   query: mocks.query,
   where: mocks.where,
 }));
-vi.mock("@/shared/firebase", () => ({ db: "db" }));
+vi.mock("@/shared/firebase", async () => ({
+  ...(await import("@/test/firebaseHelpers")).firebaseHelpers,
+  db: "db",
+}));
 vi.mock("@/features/study", () => ({
   reconcileStudySessionsWithDecks: mocks.reconcileStudySessions,
 }));
