+++ src/pages/deck-start/ui/DeckStartPage.tsx
@@ -14,7 +14,7 @@ import { usePreferences } from "@/entities/preferences";
 import { RemoteReadBoundary } from "@/shared/ui/remote-read-boundary";
 import { RouteFeedback } from "@/shared/ui/route-feedback";
 import { AppLayout } from "@/widgets/app-layout";
-import { toRemoteById } from "@/shared/api";
+import { toRemoteById } from "@/shared/firebase";
 
 import { DeckStartView } from "./DeckStartView";
