--- src/shared/api/index.ts
@@ -1,2 +1,0 @@
-export { toRemoteById } from "./remoteSnapshot";
-export type { RemoteSyncStatus } from "./remoteSnapshot";
