+++ src/shared/firebase/index.ts
@@ -8,6 +8,15 @@ import { initializeApp } from "firebase/app";
 import { connectAuthEmulator, getAuth } from "firebase/auth";
 import { connectFirestoreEmulator, initializeFirestore, persistentLocalCache } from "firebase/firestore";
 
+export {
+  firestoreTimestampDateSchema,
+  getTimestamp,
+  omitUndefined,
+  parseFirestoreDocument,
+} from "./firestoreDocument";
+export { toRemoteById } from "./remoteSnapshot";
+export type { RemoteSyncStatus } from "./remoteSnapshot";
+
 const projectId = import.meta.env.VITE_PROJECT_ID;
 const apiKey = import.meta.env.VITE_WEB_API_KEY;
