+++ src/shared/firebase/remoteSnapshot.ts
