+++ test/integration/firestore/deck.spec.ts
@@ -9,18 +9,16 @@ import "@/test/initializeTestFirestore";
 import { expect, it, describe, vi, beforeEach, type Mock } from "vitest";
 import { doc, getDoc, getFirestore } from "firebase/firestore";
 import { createCard as createCardCommand } from "@/entities/card";
-import { getTimestamp } from "@/shared/firestore";
+import { getTimestamp } from "@/shared/firebase";
 import * as UUID from "uuid";
 import { createCard, createDeck as createDeckFixture } from "@/test/factories";
 
 const uuid = UUID.v4;
 
-vi.mock("@/shared/firestore", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("@/shared/firestore")>()),
-  getTimestamp: vi.fn(),
-}));
 vi.mock("@/shared/firebase", async () => ({
+  ...(await import("@/test/firebaseHelpers")).firebaseHelpers,
   db: (await import("@/test/initializeTestFirestore")).testDb,
+  getTimestamp: vi.fn(),
 }));
 
 describe.concurrent("firestore/deck", { retry: 3 }, () => {
