+++ test/integration/firestore/subscriptions.spec.ts
@@ -16,12 +16,10 @@ import { cardStore } from "@/entities/card/model/store";
 import { deckStore } from "@/entities/deck/model/store";
 import { createCard, createDeck as createDeckFixture } from "@/test/factories";
 
-vi.mock("@/shared/firestore", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("@/shared/firestore")>()),
-  getTimestamp: vi.fn(() => 100),
-}));
 vi.mock("@/shared/firebase", async () => ({
+  ...(await import("@/test/firebaseHelpers")).firebaseHelpers,
   db: (await import("@/test/initializeTestFirestore")).testDb,
+  getTimestamp: vi.fn(() => 100),
 }));
 
 describe("Query realtime subscriptions", () => {
